Authorization added

+++ app/controllers/comments_controller.rb
@@ -28,6 +28,9 @@ class CommentsController < ApplicationController
 
   def update
     comment = Comment.find(params[:id])
+
+    authorize comment
+
     if comment.update(
       time: params[:time] || comment.time,
       author: params[:author] || comment.author,
@@ -41,6 +44,9 @@ class CommentsController < ApplicationController
 
   def destroy
     comment = Comment.find(params[:id])
+
+    authorize comment
+    
     comment.destroy
     render json: { message: 'Comment removed from Post' }, status: :ok
   end
